+++ library/include/rocsolver/rocsolver.h
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (c) 2019-2020 Advanced Micro Devices, Inc.
+ * Copyright (c) 2019-2022 Advanced Micro Devices, Inc.
  * ************************************************************************ */
 
 /*! \file
@@ -9,11 +9,12 @@
 #ifndef _ROCSOLVER_H_
 #define _ROCSOLVER_H_
 
-/* library headers */
-#include "rocsolver-aliases.h"
-#include "rocsolver-export.h"
-#include "rocsolver-extra-types.h"
-#include "rocsolver-functions.h"
-#include "rocsolver-version.h"
+/* clang-format off */
+#include "internal/rocsolver-export.h"
+#include "internal/rocsolver-extra-types.h"
+#include "internal/rocsolver-aliases.h"
+#include "internal/rocsolver-functions.h"
+#include "internal/rocsolver-version.h"
+/* clang-format on */
 
 #endif /* _ROCSOLVER_H_ */